# Reduce Engine references in expressions