Reduces possibility for hidden engine references and makes expressions more static, allowing later caching too.

## Jint.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|3d-cube|219.76 ms|0.612 ms|49017.03 KB|
| **New** |	|	| **216.89 ms (-1%)** | **0.380 ms** | **49017.87 KB (0%)** |
| Old |Run|3d-morph|168.84 ms|1.767 ms|48367.71 KB|
| **New** |	|	| **162.61 ms (-4%)** | **0.237 ms** | **48368.9 KB (0%)** |
| Old |Run|3d-raytrace|167.10 ms|0.189 ms|88376.8 KB|
| **New** |	|	| **164.49 ms (-2%)** | **0.369 ms** | **88376.7 KB (0%)** |
| Old |Run|access-binary-trees|75.98 ms|0.158 ms|60932.51 KB|
| **New** |	|	| **77.33 ms (+2%)** | **0.207 ms** | **60933.23 KB (0%)** |
| Old |Run|access-fannkuch|472.85 ms|1.516 ms|136.34 KB|
| **New** |	|	| **458.60 ms (-3%)** | **3.567 ms** | **136.34 KB (0%)** |
| Old |Run|access-nbody|190.42 ms|0.484 ms|63665.8 KB|
| **New** |	|	| **194.87 ms (+2%)** | **0.341 ms** | **63665.94 KB (0%)** |
| Old |Run|access-nsieve|169.97 ms|0.233 ms|21517.26 KB|
| **New** |	|	| **160.64 ms (-5%)** | **0.231 ms** | **21518.16 KB (0%)** |
| Old |Run|bitop(...)-byte [24]|149.14 ms|0.263 ms|56949.13 KB|
| **New** |	|	| **145.16 ms (-3%)** | **0.357 ms** | **56951.32 KB (0%)** |
| Old |Run|bitops-bits-in-byte|234.09 ms|1.196 ms|37044.95 KB|
| **New** |	|	| **246.40 ms (+5%)** | **0.438 ms** | **37047.82 KB (0%)** |
| Old |Run|bitops-bitwise-and|147.59 ms|0.742 ms|55938.9 KB|
| **New** |	|	| **146.68 ms (-1%)** | **1.767 ms** | **55938.9 KB (0%)** |
| Old |Run|bitops-nsieve-bits|220.17 ms|0.160 ms|54072.38 KB|
| **New** |	|	| **215.08 ms (-2%)** | **0.468 ms** | **54072.38 KB (0%)** |
| Old |Run|contr(...)rsive [21]|102.95 ms|0.163 ms|83184.26 KB|
| **New** |	|	| **100.32 ms (-3%)** | **0.168 ms** | **83184.26 KB (0%)** |
| Old |Run|crypto-aes|144.64 ms|0.159 ms|14010.22 KB|
| **New** |	|	| **141.89 ms (-2%)** | **0.168 ms** | **14000.9 KB (0%)** |
| Old |Run|crypto-md5|96.32 ms|0.174 ms|77906.74 KB|
| **New** |	|	| **101.50 ms (+5%)** | **0.128 ms** | **77906.77 KB (0%)** |
| Old |Run|crypto-sha1|102.81 ms|0.192 ms|64567.48 KB|
| **New** |	|	| **103.70 ms (+1%)** | **0.818 ms** | **64567.56 KB (0%)** |
| Old |Run|date-format-tofte|101.39 ms|0.342 ms|43344.18 KB|
| **New** |	|	| **103.05 ms (+2%)** | **0.195 ms** | **43344.33 KB (0%)** |
| Old |Run|date-format-xparb|56.35 ms|0.060 ms|24847.7 KB|
| **New** |	|	| **56.18 ms (0%)** | **0.064 ms** | **24847.73 KB (0%)** |
| Old |Run|math-cordic|323.15 ms|1.176 ms|81977.11 KB|
| **New** |	|	| **331.94 ms (+3%)** | **1.141 ms** | **81975.49 KB (0%)** |
| Old |Run|math-partial-sums|129.05 ms|0.393 ms|50095.98 KB|
| **New** |	|	| **124.07 ms (-4%)** | **0.282 ms** | **50095.95 KB (0%)** |
| Old |Run|math-spectral-norm|132.25 ms|0.267 ms|51833.87 KB|
| **New** |	|	| **120.76 ms (-9%)** | **0.511 ms** | **51833.84 KB (0%)** |
| Old |Run|regexp-dna|107.50 ms|0.413 ms|17666.29 KB|
| **New** |	|	| **106.14 ms (-1%)** | **0.470 ms** | **17667.31 KB (0%)** |
| Old |Run|string-base64|103.53 ms|0.185 ms|11243.06 KB|
| **New** |	|	| **101.81 ms (-2%)** | **0.316 ms** | **11242.54 KB (0%)** |
| Old |Run|string-fasta|192.31 ms|0.418 ms|148053.54 KB|
| **New** |	|	| **198.17 ms (+3%)** | **0.646 ms** | **148053.38 KB (0%)** |
| Old |Run|string-tagcloud|70.16 ms|0.836 ms|44033.92 KB|
| **New** |	|	| **70.30 ms (0%)** | **0.936 ms** | **44033.83 KB (0%)** |
| Old |Run|string-unpack-code|69.08 ms|0.157 ms|74420.88 KB|
| **New** |	|	| **71.65 ms (+4%)** | **0.129 ms** | **74420.91 KB (0%)** |
| Old |Run|strin(...)input [21]|80.86 ms|0.142 ms|39746.38 KB|
| **New** |	|	| **81.21 ms (0%)** | **0.126 ms** | **39745.33 KB (0%)** |